replace Filter struct to models
