+++ pkg/triage/triage.go
@@ -16,6 +16,7 @@ package triage
 
 import (
 	"fmt"
+	"github.com/google/triage-party/pkg/models"
 	"io"
 	"io/ioutil"
 	"time"
@@ -151,7 +152,7 @@ func (p *Party) Load(r io.Reader) error {
 }
 
 // closedAge returns how old we need to look back for a set of filters
-func closedAge(fs []hubbub.Filter) time.Duration {
+func closedAge(fs []models.Filter) time.Duration {
 	oldest := time.Duration(0)
 	if !hubbub.NeedsClosed(fs) {
 		return oldest
@@ -248,7 +249,7 @@ func processRules(raw map[string]Rule) (map[string]Rule, error) {
 
 	for id, t := range raw {
 		rules[id] = t
-		newfs := []hubbub.Filter{}
+		newfs := []models.Filter{}
 
 		for _, f := range raw[id].Filters {
 			if f.RawLabel != "" {
